Update API page and assets, add API card type toggle buttons and functionality, refine API card JSON data, add styling to cards and page
